+++ spotinst/ocean_ecs/fields_spotinst_ocean_ecs.go
@@ -150,8 +150,8 @@ func Setup(fieldsMap map[commons.FieldName]*commons.GenericField) {
 		func(resourceObject interface{}, resourceData *schema.ResourceData, meta interface{}) error {
 			clusterWrapper := resourceObject.(*commons.ECSClusterWrapper)
 			cluster := clusterWrapper.GetECSCluster()
-			if v, ok := resourceData.GetOk(string(MaxSize)); ok {
-				cluster.Capacity.SetMaximum(spotinst.Int(v.(int)))
+			if v, ok := resourceData.Get(string(MaxSize)).(int); ok && v >= 0 {
+				cluster.Capacity.SetMaximum(spotinst.Int(v))
 			}
 			return nil
 		},
@@ -189,8 +189,8 @@ func Setup(fieldsMap map[commons.FieldName]*commons.GenericField) {
 		func(resourceObject interface{}, resourceData *schema.ResourceData, meta interface{}) error {
 			clusterWrapper := resourceObject.(*commons.ECSClusterWrapper)
 			cluster := clusterWrapper.GetECSCluster()
-			if v, ok := resourceData.GetOk(string(MinSize)); ok {
-				cluster.Capacity.SetMinimum(spotinst.Int(v.(int)))
+			if v, ok := resourceData.Get(string(MinSize)).(int); ok && v >= 0 {
+				cluster.Capacity.SetMinimum(spotinst.Int(v))
 			}
 			return nil
 		},
@@ -228,8 +228,8 @@ func Setup(fieldsMap map[commons.FieldName]*commons.GenericField) {
 		func(resourceObject interface{}, resourceData *schema.ResourceData, meta interface{}) error {
 			clusterWrapper := resourceObject.(*commons.ECSClusterWrapper)
 			cluster := clusterWrapper.GetECSCluster()
-			if v, ok := resourceData.GetOk(string(DesiredCapacity)); ok {
-				cluster.Capacity.SetTarget(spotinst.Int(v.(int)))
+			if v, ok := resourceData.Get(string(DesiredCapacity)).(int); ok && v >= 0 {
+				cluster.Capacity.SetTarget(spotinst.Int(v))
 			}
 			return nil
 		},